refactor the Navbar functional component out of the header for separation of concerns
